Python: based AI Python interface on new common Python interface
