+++ python/AI/PythonAI.cpp
@@ -9,8 +9,10 @@
 #include "../../Empire/Empire.h"
 #include "../../Empire/EmpireManager.h"
 #include "../../Empire/Diplomacy.h"
+#include "../CommonFramework.h"
 #include "../PythonSetWrapper.h"
 #include "../PythonWrappers.h"
+#include "AIWrapper.h"
 
 #include <boost/python.hpp>
 #include <boost/python/suite/indexing/vector_indexing_suite.hpp>
@@ -21,6 +23,7 @@
 #include <boost/python/list.hpp>
 #include <boost/python/extract.hpp>
 #include <boost/python/scope.hpp>
+#include <boost/filesystem.hpp>
 
 using boost::python::class_;
 using boost::python::def;
@@ -41,152 +44,20 @@ using boost::python::dict;
 using boost::python::list;
 using boost::python::extract;
 
-
-////////////////////////
-// Python AIInterface //
-////////////////////////
-// disambiguation of overloaded functions
-const std::string&  (*AIIntPlayerNameVoid)(void) =              &AIInterface::PlayerName;
-const std::string&  (*AIIntPlayerNameInt)(int) =                &AIInterface::PlayerName;
-
-const Empire*       (*AIIntGetEmpireVoid)(void) =               &AIInterface::GetEmpire;
-const Empire*       (*AIIntGetEmpireInt)(int) =                 &AIInterface::GetEmpire;
-
-int                 (*AIIntNewFleet)(const std::string&, int) = &AIInterface::IssueNewFleetOrder;
-int                 (*AIIntScrap)(int)                        = &AIInterface::IssueScrapOrder;
+namespace fs = boost::filesystem;
 
 
-
-namespace {
-    // static s_save_state_string, getter and setter to be exposed to Python
-    static std::string s_save_state_string("");
-
-    static const std::string& GetStaticSaveStateString()
-    { return s_save_state_string; }
-
-    static void SetStaticSaveStateString(const std::string& new_state_string)
-    { s_save_state_string = new_state_string; }
-
-    int IssueCreateShipDesignOrderWrapper(const std::string& name, const std::string& description,
-                                          const std::string& hull, boost::python::list partsList,
-                                          const std::string& icon, const std::string& model, bool nameDescInStringTable)
-    {
-        std::vector<std::string> parts;
-        int const numParts = boost::python::len(partsList);
-        for (int i = 0; i < numParts; i++)
-            parts.push_back(boost::python::extract<std::string>(partsList[i]));
-        int result = AIInterface::IssueCreateShipDesignOrder(name, description, hull, parts, icon, model, nameDescInStringTable);
-        return result;
-    }
-
-    boost::python::list GetUserStringList(const std::string& list_key) {
-        std::list<std::string> retval;
-        UserStringList(list_key, retval);
-        boost::python::list ret_list;
-        for (std::list< std::string >::iterator it = retval.begin(); it != retval.end(); it++)
-            ret_list.append(*it);
-        return ret_list;
-    }
-}
-
-
-
-// Create the freeOrionLogger Python module, which exposes debug and error (stdout and stderr respectively)
-// sinks so Python text output can be recovered and saved in c++.  These can be accessed from within Python
-// by freeOrionLogger.log(stringParam) and freeOrionLogger.error(stringParam)
-BOOST_PYTHON_MODULE(freeOrionLogger)
-{
-    FreeOrionPython::WrapLogger();
-}
-
-/** Expose AIInterface and all associated classes to Python.
- *
- * CallPolicies:
- *
- * return_value_policy<copy_const_reference>        when returning a relatively small object, such as a string,
- *                                                  that is returned by const reference or pointer
- *
- * return_value_policy<return_by_value>             when returning either a simple data type or a temporary object
- *                                                  in a function that will go out of scope after being returned
- *
- * return_value_policy<reference_existing_object>   when returning an object from a non-member function
- */
 BOOST_PYTHON_MODULE(freeOrionAIInterface)
 {
     ///////////////////
     //  AIInterface  //
     ///////////////////
-    def("playerName",               AIIntPlayerNameVoid,            return_value_policy<copy_const_reference>());
-    def("playerName",               AIIntPlayerNameInt,             return_value_policy<copy_const_reference>());
-
-    def("playerID",                 AIInterface::PlayerID);
-    def("empirePlayerID",           AIInterface::EmpirePlayerID);
-    def("allPlayerIDs",             AIInterface::AllPlayerIDs,      return_value_policy<return_by_value>());
-
-    def("playerIsAI",               AIInterface::PlayerIsAI);
-    def("playerIsHost",             AIInterface::PlayerIsHost);
-
-    def("empireID",                 AIInterface::EmpireID);
-    def("playerEmpireID",           AIInterface::PlayerEmpireID);
-    def("allEmpireIDs",             AIInterface::AllEmpireIDs,      return_value_policy<return_by_value>());
-
-    def("getEmpire",                AIIntGetEmpireVoid,             return_value_policy<reference_existing_object>());
-    def("getEmpire",                AIIntGetEmpireInt,              return_value_policy<reference_existing_object>());
-
-    def("getUniverse",              AIInterface::GetUniverse,       return_value_policy<reference_existing_object>());
-
-    def("currentTurn",              AIInterface::CurrentTurn);
-
-    def("getAIConfigStr",           AIInterface::GetAIConfigStr,    return_value_policy<return_by_value>());
-    def("getAIDir",                 AIInterface::GetAIDir,          return_value_policy<return_by_value>());
-
-    def("initMeterEstimatesDiscrepancies",      AIInterface::InitMeterEstimatesAndDiscrepancies);
-    def("updateMeterEstimates",                 AIInterface::UpdateMeterEstimates);
-    def("updateResourcePools",                  AIInterface::UpdateResourcePools);
-    def("updateResearchQueue",                  AIInterface::UpdateResearchQueue);
-    def("updateProductionQueue",                AIInterface::UpdateProductionQueue);
-
-    def("issueFleetMoveOrder",                  AIInterface::IssueFleetMoveOrder);
-    def("issueRenameOrder",                     AIInterface::IssueRenameOrder);
-    def("issueScrapOrder",                      AIIntScrap);
-    def("issueNewFleetOrder",                   AIIntNewFleet);
-    def("issueFleetTransferOrder",              AIInterface::IssueFleetTransferOrder);
-    def("issueColonizeOrder",                   AIInterface::IssueColonizeOrder);
-    def("issueInvadeOrder",                     AIInterface::IssueInvadeOrder);
-    def("issueBombardOrder",                    AIInterface::IssueBombardOrder);
-    def("issueAggressionOrder",                 AIInterface::IssueAggressionOrder);
-    def("issueGiveObjectToEmpireOrder",         AIInterface::IssueGiveObjectToEmpireOrder);
-    def("issueChangeFocusOrder",                AIInterface::IssueChangeFocusOrder);
-    def("issueEnqueueTechOrder",                AIInterface::IssueEnqueueTechOrder);
-    def("issueDequeueTechOrder",                AIInterface::IssueDequeueTechOrder);
-    def("issueEnqueueBuildingProductionOrder",  AIInterface::IssueEnqueueBuildingProductionOrder);
-    def("issueEnqueueShipProductionOrder",      AIInterface::IssueEnqueueShipProductionOrder);
-    def("issueChangeProductionQuantityOrder",   AIInterface::IssueChangeProductionQuantityOrder);
-    def("issueRequeueProductionOrder",          AIInterface::IssueRequeueProductionOrder);
-    def("issueDequeueProductionOrder",          AIInterface::IssueDequeueProductionOrder);
-    def("issueCreateShipDesignOrder",           IssueCreateShipDesignOrderWrapper);
-
-    def("sendChatMessage",          AIInterface::SendPlayerChatMessage);
-    def("sendDiplomaticMessage",    AIInterface::SendDiplomaticMessage);
-
-    def("setSaveStateString",       SetStaticSaveStateString);
-    def("getSaveStateString",       GetStaticSaveStateString,       return_value_policy<copy_const_reference>());
-
-    def("doneTurn",                 AIInterface::DoneTurn);
-    def("userString",               make_function(&UserString,          return_value_policy<copy_const_reference>()));
-    def("userStringExists",         make_function(&UserStringExists,    return_value_policy<return_by_value>()));
-    def("userStringList",           &GetUserStringList);
-
-    def("getGalaxySetupData",       AIInterface::GetGalaxySetupData,    return_value_policy<copy_const_reference>());
-
-    boost::python::scope().attr("INVALID_GAME_TURN") = INVALID_GAME_TURN;
-
+    FreeOrionPython::WrapAI();
 
     //////////////////
     //    Empire    //
     //////////////////
     FreeOrionPython::WrapEmpire();
-
 
     /////////////////////
     // UniverseClasses //
@@ -194,12 +65,10 @@ BOOST_PYTHON_MODULE(freeOrionAIInterface)
     FreeOrionPython::WrapUniverseClasses();
     FreeOrionPython::WrapGalaxySetupData();
 
-
     ///////////////////
     //     Enums     //
     ///////////////////
     FreeOrionPython::WrapGameStateEnums();
-
 
     ////////////////////
     // STL Containers //
@@ -222,108 +91,57 @@ BOOST_PYTHON_MODULE(freeOrionAIInterface)
 //////////////////////
 //     PythonAI     //
 //////////////////////
-static dict         s_main_namespace = dict();
-static object       s_ai_module = object();
-static PythonAI*    s_ai = 0;
-#if defined(FREEORION_MACOSX) || defined(FREEORION_WIN32)
-static char         s_python_home[1024];
-static char         s_python_program_name[1024];
-#endif
-PythonAI::PythonAI() {
-    DebugLogger() << "PythonAI::PythonAI()";
-    // in order to expose a getter for it to Python, s_save_state_string must be static, and not a member
-    // variable of class PythonAI, because the exposing is done outside the PythonAI class and there is no
-    // access to a pointer to PythonAI
-    if (s_ai)
-        throw std::runtime_error("Attempted to create more than one Python AI instance");
-
-    s_ai = this;
-
-    try {
-#if defined(FREEORION_MACOSX) || defined(FREEORION_WIN32)
-        // There have been recurring issues on Windows and OSX to get FO to use the
-        // Python framework shipped with the app (instead of falling back on the ones
-        // provided by the system). These API calls have been added in an attempt to
-        // solve the problems. Not sure if they are really required, but better save
-        // than sorry... ;)
-        strcpy(s_python_home, GetPythonHome().string().c_str());
-        Py_SetPythonHome(s_python_home);
-        DebugLogger() << "Python home set to " << Py_GetPythonHome();
-        strcpy(s_python_program_name, (GetPythonHome() / "Python").string().c_str());
-        Py_SetProgramName(s_python_program_name);
-        DebugLogger() << "Python program name set to " << Py_GetProgramFullPath();
-#endif
-        Py_Initialize();                // initializes Python interpreter, allowing Python functions to be called from C++
-        DebugLogger() << "Python initialized";
-
-        DebugLogger() << "Python version: " << Py_GetVersion();
-        DebugLogger() << "Python prefix: " << Py_GetPrefix();
-        DebugLogger() << "Python module search path: " << Py_GetPath();
-
-        DebugLogger() << "Initializing C++ interfaces for Python";
-
-        initfreeOrionLogger();          // allows the "freeOrionLogger" C++ module to be imported within Python code
-        initfreeOrionAIInterface();     // allows the "freeOrionAIInterface" C++ module to be imported within Python code
-    } catch (...) {
-        ErrorLogger() << "Unable to initialize Python interpreter.";
-        return;
+bool PythonAI::Initialize() {
+    if (PythonBase::Initialize()) {
+        try {
+            object initAIPythonFunction = m_python_module_ai.attr("initFreeOrionAI");
+            initAIPythonFunction();
+        } catch (error_already_set err) {
+            PyErr_Print();
+            return false;
+        }
+        return true;
     }
-
-    try {
-        // get main namespace, needed to run other interpreted code
-        object main_module = import("__main__");
-        s_main_namespace = extract<dict>(main_module.attr("__dict__"));
-    } catch (error_already_set err) {
-        ErrorLogger() << "Unable to set up main namespace in Python.";
-        PyErr_Print();
-        return;
-    }
-
-    try {
-        // set up Logging by redirecting stdout and stderr to exposed logging functions
-        std::string logger_script = "import sys\n"
-                                    "import freeOrionLogger\n"
-                                    "class debugLogger:\n"
-                                    "  def write(self, stng):\n"
-                                    "    freeOrionLogger.log(stng)\n"
-                                    "class errorLogger:\n"
-                                    "  def write(self, stng):\n"
-                                    "    freeOrionLogger.error(stng)\n"
-                                    "sys.stdout = debugLogger()\n"
-                                    "sys.stderr = errorLogger()\n"
-                                    "print ('Python stdout and stderr redirected')";
-        object ignored = exec(logger_script.c_str(), s_main_namespace, s_main_namespace);
-    } catch (error_already_set err) {
-        ErrorLogger() << "Unable to redirect Python stdout and stderr.";
-        return;
-    }
-
-    try {
-        // tell Python the path in which to locate AI script file
-        std::string AI_path = (GetResourceDir() / GetOptionsDB().Get<std::string>("ai-path")).string();
-        std::string path_command = "sys.path.append(r'" + AI_path + "')";
-        object ignored = exec(path_command.c_str(), s_main_namespace, s_main_namespace);
-
-        // import AI script file and run initialization function
-        s_ai_module = import("FreeOrionAI");
-        object initAIPythonFunction = s_ai_module.attr("initFreeOrionAI");
-        initAIPythonFunction();
-
-        //ignored = exec(fo_interface_import_script.c_str(), s_main_namespace, s_main_namespace);
-    } catch (error_already_set err) {
-        PyErr_Print();
-        return;
-    }
-
-    DebugLogger() << "Initialized Python AI";
+    else
+        return false;
 }
 
-PythonAI::~PythonAI() {
-    DebugLogger() << "Cleaning up / destructing Python AI";
-    Py_Finalize();      // stops Python interpreter and release its resources
-    s_ai = 0;
-    s_main_namespace = dict();
-    s_ai_module = object();
+bool PythonAI::InitModules() {
+    DebugLogger() << "Initializing AI Python modules";
+
+    // allows the "freeOrionAIInterface" C++ module to be imported within Python code
+    try {
+        initfreeOrionAIInterface();
+    } catch (...) {
+        ErrorLogger() << "Unable to initialize 'freeOrionAIInterface' AI Python module";
+        return false;
+    }
+
+    // Confirm existence of the directory containing the AI Python scripts
+    // and add it to Pythons sys.path to make sure Python will find our scripts
+    std::string ai_path = (GetResourceDir() / GetOptionsDB().Get<std::string>("ai-path")).string();
+    DebugLogger() << "AI Python script path: " << ai_path;
+    if (!fs::exists(ai_path)) {
+        ErrorLogger() << "Can't find folder containing AI scripts";
+        return false;
+    }
+    if (!AddToSysPath(ai_path)) {
+        ErrorLogger() << "Can't add folder containing AI scripts to sys.path";
+        return false;
+    }
+
+    try {
+        // import universe generator script file
+        m_python_module_ai = import("FreeOrionAI");
+    }
+    catch (error_already_set err) {
+        ErrorLogger() << "Unable to import AI script";
+        PyErr_Print();
+        return false;
+    }
+
+    DebugLogger() << "AI Python modules successfully initialized!";
+    return true;
 }
 
 void PythonAI::GenerateOrders() {
@@ -334,7 +152,7 @@ void PythonAI::GenerateOrders() {
     try {
         // call Python function that generates orders for current turn
         //DebugLogger() << "PythonAI::GenerateOrders : getting generate orders object";
-        object generateOrdersPythonFunction = s_ai_module.attr("generateOrders");
+        object generateOrdersPythonFunction = m_python_module_ai.attr("generateOrders");
         //DebugLogger() << "PythonAI::GenerateOrders : generating orders";
         generateOrdersPythonFunction();
     } catch (error_already_set err) {
@@ -349,7 +167,7 @@ void PythonAI::GenerateOrders() {
 void PythonAI::HandleChatMessage(int sender_id, const std::string& msg) {
     try {
         // call Python function that responds or ignores a chat message
-        object handleChatMessagePythonFunction = s_ai_module.attr("handleChatMessage");
+        object handleChatMessagePythonFunction = m_python_module_ai.attr("handleChatMessage");
         handleChatMessagePythonFunction(sender_id, msg);
     } catch (error_already_set err) {
         PyErr_Print();
@@ -359,7 +177,7 @@ void PythonAI::HandleChatMessage(int sender_id, const std::string& msg) {
 void PythonAI::HandleDiplomaticMessage(const DiplomaticMessage& msg) {
     try {
         // call Python function to inform of diplomatic message change
-        object handleDiplomaticMessagePythonFunction = s_ai_module.attr("handleDiplomaticMessage");
+        object handleDiplomaticMessagePythonFunction = m_python_module_ai.attr("handleDiplomaticMessage");
         handleDiplomaticMessagePythonFunction(msg);
     } catch (error_already_set err) {
         PyErr_Print();
@@ -369,7 +187,7 @@ void PythonAI::HandleDiplomaticMessage(const DiplomaticMessage& msg) {
 void PythonAI::HandleDiplomaticStatusUpdate(const DiplomaticStatusUpdateInfo& u) {
     try {
         // call Python function to inform of diplomatic status update
-        object handleDiplomaticStatusUpdatePythonFunction = s_ai_module.attr("handleDiplomaticStatusUpdate");
+        object handleDiplomaticStatusUpdatePythonFunction = m_python_module_ai.attr("handleDiplomaticStatusUpdate");
         handleDiplomaticStatusUpdatePythonFunction(u);
     } catch (error_already_set err) {
         PyErr_Print();
@@ -377,10 +195,10 @@ void PythonAI::HandleDiplomaticStatusUpdate(const DiplomaticStatusUpdateInfo& u)
 }
 
 void PythonAI::StartNewGame() {
-    s_save_state_string = "";
+    FreeOrionPython::ClearStaticSaveStateString();
     try {
         // call Python function that sets up the AI to be able to generate orders for a new game
-        object startNewGamePythonFunction = s_ai_module.attr("startNewGame");
+        object startNewGamePythonFunction = m_python_module_ai.attr("startNewGame");
         startNewGamePythonFunction(m_aggression);
     } catch (error_already_set err) {
         PyErr_Print();
@@ -389,11 +207,11 @@ void PythonAI::StartNewGame() {
 
 void PythonAI::ResumeLoadedGame(const std::string& save_state_string) {
     //DebugLogger() << "PythonAI::ResumeLoadedGame(" << save_state_string << ")";
-    s_save_state_string = save_state_string;
+    FreeOrionPython::SetStaticSaveStateString(save_state_string);
     try {
         // call Python function that deals with the new state string sent by the server
-        object resumeLoadedGamePythonFunction = s_ai_module.attr("resumeLoadedGame");
-        resumeLoadedGamePythonFunction(s_save_state_string);
+        object resumeLoadedGamePythonFunction = m_python_module_ai.attr("resumeLoadedGame");
+        resumeLoadedGamePythonFunction(FreeOrionPython::GetStaticSaveStateString());
     } catch (error_already_set err) {
         PyErr_Print();
     }
@@ -403,11 +221,11 @@ const std::string& PythonAI::GetSaveStateString() {
     try {
         // call Python function that serializes AI state for storage in save file and sets s_save_state_string
         // to contain that string
-        object prepareForSavePythonFunction = s_ai_module.attr("prepareForSave");
+        object prepareForSavePythonFunction = m_python_module_ai.attr("prepareForSave");
         prepareForSavePythonFunction();
     } catch (error_already_set err) {
         PyErr_Print();
     }
     //DebugLogger() << "PythonAI::GetSaveStateString() returning: " << s_save_state_string;
-    return s_save_state_string;
+    return FreeOrionPython::GetStaticSaveStateString();
 }
